outsource test_traditions object to a shared variables file
